Break infra use_cases import cycle
