+++ apps/server/tests/integration/test_review_guardrails_core_regressions.py
@@ -15,9 +15,9 @@ import pytest
 from vibesensor.infra.runtime.registry import _sanitize_name
 from vibesensor.infra.workers.worker_pool import WorkerPool
 from vibesensor.shared.json_utils import as_float_or_none, as_int_or_none
+from vibesensor.shared.order_bands import build_order_bands
 from vibesensor.shared.sampling import bounded_sample
 from vibesensor.shared.types.backend_types import new_car_id
-from vibesensor.use_cases.diagnostics.order_bands import build_order_bands
 
 # ---------------------------------------------------------------------------
 # Item 1 + 2: Public API naming in domain_models
@@ -50,7 +50,7 @@ class TestDomainModelsPublicAPI:
 
 
 # ---------------------------------------------------------------------------
-# Item 4: build_order_bands lives in order_bands
+# Item 4: build_order_bands lives in shared.order_bands
 # ---------------------------------------------------------------------------
