+++ apps/server/tests/use_cases/diagnostics/test_common_vibration_causes.py
@@ -5,7 +5,7 @@ from math import pi
 import pytest
 
 from vibesensor.shared.constants import KMH_TO_MPS
-from vibesensor.use_cases.diagnostics.order_bands import (
+from vibesensor.shared.order_bands import (
     build_diagnostic_settings,
     vehicle_orders_hz,
 )
