+++ apps/server/tests/use_cases/diagnostics/test_single_source_of_truth.py
@@ -61,7 +61,7 @@ def test_as_float_single_source_of_truth() -> None:
     from runlog, not a local re-definition.
     """
     from vibesensor.shared.json_utils import as_float_or_none
-    from vibesensor.use_cases.diagnostics.order_bands import as_float_or_none as ob_as_float
+    from vibesensor.shared.order_bands import as_float_or_none as ob_as_float
 
     assert ob_as_float is as_float_or_none, (
         "order_bands.as_float_or_none must be imported from runlog.as_float_or_none"
