+++ apps/server/vibesensor/adapters/http/health_snapshot.py
@@ -1,11 +1,11 @@
-"""Health snapshot builder — pure business logic, no HTTP concerns."""
+"""Health snapshot payload builder for the HTTP health endpoint."""
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Literal, TypedDict
+from typing import TYPE_CHECKING, Literal
 
 from vibesensor.infra.runtime.processing_loop import ProcessingHealth
-from vibesensor.use_cases.run.logger import RunRecorderHealthSnapshot
+from vibesensor.shared.types.health_snapshot import HealthSnapshotData
 
 if TYPE_CHECKING:
     from vibesensor.infra.processing import SignalProcessor
@@ -13,32 +13,6 @@ if TYPE_CHECKING:
     from vibesensor.infra.runtime.processing_loop import ProcessingLoopState
     from vibesensor.infra.runtime.registry import ClientRegistry
     from vibesensor.use_cases.run import RunRecorder
-
-
-class HealthSnapshotData(TypedDict):
-    """Typed snapshot returned by :func:`build_health_snapshot`."""
-
-    status: Literal["ok", "warn", "degraded"]
-    startup_state: str
-    startup_phase: str
-    startup_error: str | None
-    startup_warnings: list[str]
-    background_task_failures: dict[str, str]
-    processing_state: str
-    processing_failures: int
-    processing_failure_categories: dict[str, int]
-    processing_last_failure: str | None
-    sample_rate_mismatch_count: int
-    frame_size_mismatch_count: int
-    degradation_reasons: list[str]
-    data_loss: dict[str, int]
-    persistence: RunRecorderHealthSnapshot
-    intake_stats: dict[str, object]
-    tick_duration_s: float | None
-    max_tick_duration_s: float | None
-    tick_count: int
-    db_last_write_duration_s: float | None
-    db_max_write_duration_s: float | None
 
 
 def build_health_snapshot(
@@ -96,10 +70,9 @@ def build_health_snapshot(
         degradation_reasons.append("analyzing_runs_present")
     if persistence["last_completed_run_error"]:
         degradation_reasons.append("last_analysis_failed")
+    status: Literal["ok", "warn", "degraded"] = "ok"
     if degradation_reasons:
         status = "degraded" if has_error else "warn"
-    else:
-        status = "ok"
     return {
         "status": status,
         "startup_state": health_state.startup_state,
