+++ apps/server/vibesensor/infra/runtime/__init__.py
@@ -2,7 +2,6 @@
 
 Submodules
 ----------
-- ``state``: top-level runtime assembly (RuntimeState)
 - ``builders``: focused service builders
 - ``processing_loop``: ProcessingLoop (async tick loop + failure tracking)
 - ``ws_broadcast``: WsBroadcastService (payload assembly + cache)
@@ -12,10 +11,8 @@ Submodules
 
 from vibesensor.infra.runtime.health_state import RuntimeHealthState
 from vibesensor.infra.runtime.processing_loop import ProcessingLoopState
-from vibesensor.infra.runtime.state import RuntimeState
 
 __all__ = [
     "ProcessingLoopState",
     "RuntimeHealthState",
-    "RuntimeState",
 ]
