+++ apps/server/vibesensor/infra/runtime/rotational_speeds.py
@@ -7,11 +7,11 @@ from typing import TYPE_CHECKING
 
 from vibesensor.domain.speed_source import SpeedSource
 from vibesensor.shared.constants import SECONDS_PER_MINUTE
+from vibesensor.shared.order_bands import build_order_bands, vehicle_orders_hz
 from vibesensor.shared.types.payload_types import (
     RotationalSpeedsPayload,
     RotationalSpeedValuePayload,
 )
-from vibesensor.use_cases.diagnostics import build_order_bands, vehicle_orders_hz
 
 if TYPE_CHECKING:
     from vibesensor.adapters.gps.gps_speed import GPSSpeedMonitor
